+++ packages/vue-stack/src/constants.ts
@@ -1,0 +1,1 @@
+export const V_STACK_ACTIVATED_ATTR = 'data-v-stack-activated';